+++ MagicEyes/src/backend/net/net_watcher/include/net_watcher.h
@@ -56,10 +56,16 @@ typedef unsigned long long u64;
 #define MAX_EVENTS 1024
 #define CACHEMAXSIZE 5
 #define PID 32
-#define NS(x) ((x) * 1000000000) // 1 秒等于 10^9 纳秒
+#define NS(x) ((x) * 1000000000LL) // 1 秒等于 10^9 纳秒
 #define TIME_THRESHOLD_NS NS(10) 
 
 typedef u64 stack_trace_t[MAX_STACK_DEPTH];
+
+typedef struct
+{
+    char key[256];
+    u32 value;
+} kv_pair;
 
 struct conn_t
 {
@@ -316,6 +322,5 @@ struct tcp_rate
 struct tcp_args_s {
     u64 sample_period; 
 };
-
 
 #endif /* __NET_WATCHER_H */
